togglePopUpMenu module
- Moved toggleUserMenu() from header.js to a separate module on
togglePopUpMenu().
